This commit was manufactured by cvs2svn to create tag 'Ext-5-1-3-pre1'.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/tags/Ext-5-1-3-pre1@12375 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/agent/nsDebug.c
@@ -108,7 +108,7 @@ handle_nsDebugEnabled(netsnmp_mib_handler *handler,
                 netsnmp_agent_request_info *reqinfo,
                 netsnmp_request_info *requests)
 {
-    int enabled;
+    long enabled;
     netsnmp_request_info *request=NULL;
 
     switch (reqinfo->mode) {
@@ -163,7 +163,7 @@ handle_nsDebugOutputAll(netsnmp_mib_handler *handler,
                 netsnmp_agent_request_info *reqinfo,
                 netsnmp_request_info *requests)
 {
-    int enabled;
+    long enabled;
     netsnmp_request_info *request=NULL;
 
     switch (reqinfo->mode) {
@@ -218,7 +218,7 @@ handle_nsDebugDumpPdu(netsnmp_mib_handler *handler,
                 netsnmp_agent_request_info *reqinfo,
                 netsnmp_request_info *requests)
 {
-    int enabled;
+    long enabled;
     netsnmp_request_info *request=NULL;
 
     switch (reqinfo->mode) {
